% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pythag.R
\name{pythag}
\alias{pythag}
\title{Calculate the length of the third side of a right triangle.}
\usage{
pythag(a, b, c = NULL)
}
\arguments{
\item{a}{The length of one of the two shorter sides of the right triangle.}

\item{b}{The length of the other shorter side of the right triangle.}

\item{c}{The length of the hypotenuse (not allowed, set to NULL).}
}
\value{
The length of the third side of a right triangle.
}
\description{
This function calculates the length of the third side of a right triangle
using the Pythagorean theorem
}
\details{
This function checks various conditions:
\itemize{
\item Must provide the length of at least two sides.
\item Cannot provide the length of all three sides.
\item All arguments must be numeric.
}
}
\examples{
pythag(a = 3, b = 4)
[1] 5

}
